fix floating point issue

+++ clang/runtime/dpct-rt/include/dpct/codepin/codepin.hpp
@@ -239,7 +239,7 @@ public:
     for (size_t i = 0; i < size; ++i) {
       if (size > CODEPIN_SAMPLING_THRESHOLD && i != 0) {
         float r = static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
-        if (r > CODEPIN_SAMPLING_PERCENT/100)
+        if (r > (float)CODEPIN_SAMPLING_PERCENT/(float)100)
           continue;
       }
       auto obj = arr.object();
